+++ icat/view.py
@@ -61,6 +61,7 @@ class InteractiveView(pn.viewable.Viewer):
         self.model.anchor_list.on_anchor_changed(
             self._update_data_table_on_anchor_change
         )
+        self.model.anchor_list.on_anchor_added(self._update_data_table_on_anchor_added)
         self.model.anchor_list.on_anchor_removed(self._remove_list_anchor_from_viz)
         self.anchorviz.on_anchor_add(self._add_viz_anchor_to_list)
         self.anchorviz.observe(
@@ -235,6 +236,13 @@ class InteractiveView(pn.viewable.Viewer):
 
     def _update_data_table_on_anchor_change(self, id: str, property: str, value: Any):
         """Whenever an anchor changes, refresh the data table (this is so that any
+        keyword changes correctly update highlighting)"""
+        # NOTE: unsure if this will cause any "jumping" or lag, if so, look for a keywords
+        # property?
+        self.model.data.update_trigger = True
+
+    def _update_data_table_on_anchor_added(self, anchor: Anchor):
+        """Whenever an anchor is added, refresh the data table (this is so that any
         keyword changes correctly update highlighting)"""
         # NOTE: unsure if this will cause any "jumping" or lag, if so, look for a keywords
         # property?